Move config out of root
